split module.go file
